+++ src/components/common/Button/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Button';
+export * from './types';
